+++ saleor/graphql/product/tests/queries/test_collections_query_with_filter_and_sort.py
@@ -2,7 +2,6 @@ import datetime
 
 import graphene
 import pytest
-from graphql_relay import to_global_id
 
 from .....product.models import Collection, CollectionChannelListing
 from .....tests.utils import dummy_editorjs
@@ -19,7 +18,15 @@ from ....tests.utils import (
         ({"published": "HIDDEN"}, 1),
         ({"search": "-published1"}, 1),
         ({"search": "Collection3"}, 1),
-        ({"ids": [to_global_id("Collection", 2), to_global_id("Collection", 3)]}, 2),
+        (
+            {
+                "ids": [
+                    graphene.Node.to_global_id("Collection", 2),
+                    graphene.Node.to_global_id("Collection", 3),
+                ]
+            },
+            2,
+        ),
     ],
 )
 def test_collections_query_with_filter(